chore(rbo, codemanagement) MCOL-6143: move rbo to separate dir
