+++ tests/rewritetest.cpp
